Create memorygame.py
Memory Game is a simple game created using Python and the Tkinter library. In this game, players must match pairs of number. It assesses the player's visual memory abilities and, upon completion, displays a completion time message.

Key attributes:
Game Timer: To record the amount of time needed to finish the game, the timer will start running as soon as the game begins.
Direct Feedback: A pop-up to document errors is displayed along with instant feedback when matches are made or not.
Point System: After every pair has been located, the game is over and the total number of pairs found will be tallied.

Methods for Playing:
To begin, click the "Are you ready?" button.
Match all the pairs to complete the fame.

Technology:
Language: Python
GUI: Tkinter